requirements.txt
fastapi[standard]
langchain
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
python-dotenv
